Change conversion of NodeList to Array
There's a trick to convert a NodeList into an Array, in order to use
some Array functions on it. (in this case, forEach):

    Array.slice.call(null, document.querySelectorAll(target)).forEach(function(element) {

That works in Firefox, but not in Chrome.  The following works in both:

    Array.prototype.slice.call(document.querySelectorAll(target)).forEach(function(element) {

And more compellingly, Array.prototype.slice is included in the
standard, according to MDN.  So, if it doesn't work in your browser,
it's the browser's fault. (Which admittedly isn't everything, but I feel
I have to set my personal bar for 'good enough' somewhere short of
omniscience lest madness take me in the night.  And this seems
reasonable.)